+++ src/components/date_picker/index.ts
@@ -1,0 +1,28 @@
+export * from './super_date_picker';
+
+export { EuiDatePicker, EuiDatePickerProps } from './date_picker';
+
+export {
+  EuiDatePickerRange,
+  EuiDatePickerRangeProps,
+} from './date_picker_range';
+
+export {
+  DurationRange as EuiSuperDatePickerCommonRange,
+  DurationRange as EuiSuperDatePickerDurationRange,
+  DurationRange as EuiSuperDatePickerRecentRange,
+  TimeUnitId,
+  TimeUnitFromNowId,
+  TimeUnitLabel,
+  TimeUnitLabelPlural,
+  AbsoluteDateMode,
+  RelativeDateMode,
+  NowDateMode,
+  DateMode,
+  OnRefreshChangeProps,
+  ShortDate,
+  RelativeParts,
+  RelativeOption,
+  QuickSelect,
+  QuickSelectPanel as EuiSuperDatePickerQuickSelectPanel,
+} from './types';